www/requirements.txt
flask
flask-sqlalchemy
mysql-python
mysql-connector-python
flask-cors

